% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn_predict}
\alias{knn_predict}
\title{Predict with a KNN model}
\usage{
knn_predict(
  model,
  newx,
  k = 5,
  weights = c("uniform", "distance"),
  method = c("auto", "brute"),
  use_rcpp = TRUE
)
}
\arguments{
\item{model}{A 'knnpack_model' from knn_fit()}

\item{newx}{Numeric matrix/data.frame of new features}

\item{k}{Integer K}

\item{weights}{"uniform" or "distance"}

\item{method}{"auto" or "brute" (currently uses brute-force distances)}

\item{use_rcpp}{If TRUE, use Rcpp distance kernel when available}
}
\value{
Vector of predictions (class labels or numeric)
}
\description{
Predict with a KNN model
}
