% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn_score}
\alias{knn_score}
\title{Score a KNN model (accuracy or RMSE)}
\usage{
knn_score(model, x, y, k = 5, ...)
}
\arguments{
\item{model}{knnpack_model}

\item{x}{Matrix/data.frame of features}

\item{y}{True outcomes}

\item{k}{Number of neighbors}

\item{...}{Passed to knn_predict}
}
\value{
numeric; accuracy for classification, RMSE for regression
}
\description{
Score a KNN model (accuracy or RMSE)
}
